fix for connector issues.
